React Native: Alert

+++ ReactNative/confusion/components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from "react-native-elements";
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
@@ -35,7 +35,25 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavourite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favourite?',
+                            'Are you sure you wish to delete the favourite dish '
+                            + item.name + '?',
+                            [
+                                {
+                                    text: 'Cancel',
+                                    onPress: () => console.log(item.name + ' Not Deleted'),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text: 'Ok',
+                                    onPress: () => this.props.deleteFavourite(item.id)
+                                }
+                            ],
+                            {cancelable: false}
+                        );
+                    }
                 }
             ];
